+++ app/src/main/java/com/example/ncmonitor/MainActivity.kt
@@ -20,26 +20,12 @@ import kotlin.math.round
 
 const val GB = 1073741824
 const val Byte = 1024
-const val PERMISSION_REQUEST = 10
-
-/* your Nextcloud username */
-const val USERNAME = ""
-
-/* your Nextcloud password */
-const val PASSWORD = ""
-
-/* your Nextcloud endpoint. ?format=json is required */
-const val URL = ""
 
 /* number of cores in your server. 4 for RPi4. */
 const val N_CORES = 4
 
 
-
 class MainActivity : WearableActivity() {
-
-    var client = OkHttpClient()
-    var request = OkHttpRequest(client)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -48,31 +34,70 @@ class MainActivity : WearableActivity() {
         // Enables Always-on
         setAmbientEnabled()
 
-        checkAndRequestPermissions()
+        val extras = intent.extras
+        if (extras != null)
+            showResults(extras.getString("response")!!)
 
-        requestNcStatus()
+        //requestNcStatus()
     }
 
-    private fun checkAndRequestPermissions(){
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
 
-            val permissions = arrayOf(
-                android.Manifest.permission.INTERNET
-            )
+    private fun showResults(response: String)
+    {
+        try {
+            val json = JSONObject(response)
+            Log.d("requestNcStatus","Request Successful!!")
+            Log.d("json_response", json.toString())
+            warning_icon.visibility = View.GONE
+            ok_icon.visibility = View.VISIBLE
 
-            val permissionsDenied = arrayListOf<String>()
+            val responseObject = json.getJSONObject("ocs")
+            val metaObject = responseObject.getJSONObject("meta")
 
-            for (permission in permissions) {
-                if (checkSelfPermission(permission) == PackageManager.PERMISSION_DENIED) {
-                    permissionsDenied.add(permission)
-                }
+            val statusCode = metaObject.getInt("statuscode")
+
+            if (statusCode == 200) {
+                status_code_response_layout.visibility = View.GONE
+                status_message_response_layout.visibility = View.GONE
+
+                val dataObject =  responseObject.getJSONObject("data")
+                val nextcloudObject =  dataObject.getJSONObject("nextcloud")
+                val systemObject =  nextcloudObject.getJSONObject("system")
+
+                // cpu load
+                val cpuLoad = systemObject.getJSONArray("cpuload")[0] as Double
+                // ram
+                val ramTotal = systemObject.getLong("mem_total")
+                val ramFree = systemObject.getLong("mem_free")
+                // swap
+                val swapTotal = systemObject.getLong("swap_total")
+                val swapFree = systemObject.getLong("swap_free")
+                // disk
+                val diskFree = systemObject.getDouble("freespace")
+
+                updateViews(cpuLoad, ramTotal-ramFree, ramTotal,
+                    swapTotal-swapFree, swapTotal, diskFree)
+
+
+            } else {
+                cpu_layout.visibility = View.GONE
+                ram_layout.visibility = View.GONE
+                swap_layout.visibility = View.GONE
+                disk_layout.visibility = View.GONE
+                status_code_response_used_placeholder.text = statusCode.toString()
+                status_message_response_placeholder.text = metaObject.getString("message")
             }
 
-            if(permissionsDenied.isNotEmpty())
-                requestPermissions(permissionsDenied.toTypedArray(), PERMISSION_REQUEST)
+        } catch (e: JSONException) {
+            Log.d("requestNcStatus","JsonException!!")
+            warning_icon.visibility = View.GONE
+            error_icon.visibility = View.VISIBLE
+            e.printStackTrace()
         }
     }
 
+
+/*
     private fun requestNcStatus()
     {
         val credentials = basic(USERNAME, PASSWORD)
@@ -127,10 +152,6 @@ class MainActivity : WearableActivity() {
                         }
 
 
-
-
-
-
                         //val docs = json.getJSONArray("docs")
                         //this@MainActivity.fetchComplete()
                     } catch (e: JSONException) {
@@ -157,9 +178,11 @@ class MainActivity : WearableActivity() {
                     status_message_response_placeholder.text = e?.message
                 }
             }
+
         })
 
     }
+ */
 
     fun Double.format(digits: Int) = "%.${digits}f".format(this)
 
@@ -174,7 +197,6 @@ class MainActivity : WearableActivity() {
         ram_total_placeholder.text = (ramTotal / Byte).toString()
         swap_used_placeholder.text = (swapBusy /Byte).toString()
         swap_total_placeholder.text = (swapTotal / Byte).toString()
-
 
         disk_used_placeholder.text = (diskFree / GB).format(2)
     }