+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## 0.2.0 (July, 5, 2020)
+
+- added login activity
+- added login layout
+- progress bar during the execution of the GET
+- added NcMonitor app icon
+- moved the login and networking logic to the login activity
+
 ## 0.1.0 (July, 3, 2020)
 
-- ambient dispplay support
+- ambient display support
 - horizontal layout for error message
 - added OkHttpRequest class
 - new layout